Added different colors for enemy tanks. Fixed the movement for enemy tanks, so their rotations will act kind of the same like player.
Added a track distance and an attack distance, so the enemy tanks will track player if their distance is greater than track distance and attack player if their distance is less than attack distance. The enemy tanks will only shoot when they closely face to player.

Added a collision between enemy's bullet and player, and between player's bullet and enemy tanks. Also added a collision between the enemy's bullet and shield, and the light hit effect of the shield as well. The bullet will be detached while it hits the tanks,  shield, or is far away from player.