Added hit points and a bar for that for all the enenmies, player and shield. If the bullet hits something, the hit points will decrease. The shield will be detached if its hit points is zero. The game will be over if the hitpoints for all the enemies or the player become zero and some text will be popped up.
Added an action if there are collisions between all the tanks. The enemy tanks will be bounced back if they hit player or part of themselves. So there kind of will not exist overlapping. This is not physics. Rigidbodycontrol will significantly influence the FPS. So I delete all of the rigidBodyControl.

Also added a number for the bullet. So the number of bullet is not infinite. The number can be seen on the left up corner of the screen.